--- frontend/src/containers/MediaUpload/index.tsx
@@ -1,1 +1,0 @@
-export { UploadButton } from './MediaUpload';
