+++ frontend/src/features/MediaUpload/MediaUploadDialog.tsx
@@ -1,8 +1,8 @@
 import { Dialog, Button } from '@components';
-import { useMediaUpload } from '@hooks/useMediaUpload';
 import React, { useState } from 'react';
 
 import { FileSelect } from './FileSelect';
+import { useMediaUpload } from './useMediaUpload';
 
 import type { ContentType } from '@/client';
