+++ frontend/src/features/MediaUpload/MediaUploadMonitor.styled.tsx
