+++ frontend/src/features/MediaUpload/index.tsx
@@ -1,0 +1,4 @@
+export { UploadButton } from './UploadButton';
+export { useMediaUpload, MediaUploadProvider } from './useMediaUpload';
+export { MediaUploadMonitor } from './MediaUploadMonitor';
+
