+++ frontend/src/pages/System/Storages/Storages.tsx
@@ -1,4 +1,3 @@
-import { setPageTitle } from '@actions';
 import { Form, FormRow, InputSwitch } from '@components';
 import { Section, Spacer } from '@components';
 import { useNebula } from '@features/Nebula';
